backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
resend==2.19.0
python-dotenv==1.0.0
pillow==10.1.0
matplotlib==3.8.2
numpy==1.26.2
cairosvg==2.7.1
